remove raftstore_router.go
